Adding sort project with different sorting techniques
